Create grid and content for bottom header
Set a specific height in .header and then set specific height in
.topHeader and .bottomHeader so no overflowing can occur. Then add in
divs to index.html and set up a grid where each div can be placed in a
specific location. For profile header, ensure that the max-height is
100% so it fills the .bottomHeader div to 100%, which is 80% of the
header which is 20vh of the viewport. Width is auto so the aspect
ratio remains the same