Format bottom header with specific heights
Use flexbox in header and direction in columns to set the top and bottom
header displayed top to down. Then set a max height on top header, the
bottom header is now free to take up remaining space. Set a percentage
margin for bottom header as to not cause overflowing potential to
compensate for px's.