Create framework for body portion of website
Use display grid to separate the different sections of the body portion.
Then use headings for each followed by a container that has the small
tabs of block. Since display block is defaulted for divs, we dont need
to worry about display on event and tasks container. But for body tabs,
we align in a 2x3 grid