Create grid template for tabs
Use display: grid to create a grid template for each tab. Use fr and
make the last cell in the bottom right really small as compared to
the one with sentences so that we can place an image in the cell. Have
the image be placed in that cell using css and maximize height and
width auto