Remove style from last child div and add framework
Use the same direct child selector but have a last child selector and
set the border to zero to overwrite the border. Add framework of the
event and tasks div by making display:grid on it. This will give us
opportunity to use gaps and then have the child like EventContainer
stretch to its full potential for its 2 / 3 row.