Add content and styling to tasks container
Add some padding of 5% and then add content in index.html. We have an
image followed by a description. Then we make a display grid on the
container containing these images and divs and set two columns and four
rows. Then we will add some gap to the column. Set the image to max
height of 100 and then width auto to make it take the entire space and
then place-self center on the cell. Then for description add some basic
styling but then only align self so it centers vertically but not
horizontally