Add new football image and increase tab styling
Football image was too small so find a bigger one and have it max height
on the container so it is in the same size as all other images in the
sidebar. Then add in new columns and stretch the header and description
to all columns and add gap between columns. Then add in new pictures
that go in these cells